Assignment_4/Alex (a voice assistant system)/requirements.txt
SpeechRecognition==3.14.3
pyttsx3==2.90
PyAudio
wikipedia 
google-generativeai